All tails are now split-rendered, fixes ID Card Picture Bug
I chose to split-render all tails because I wanted to sort out an issue
where if you had a certain set of cyborg legs or were missing legs and
had a tail, it'd look demented what with there being floating pieces and
such. That is no longer an issue, you will now see a full tail even in
the absence of legs. When facing south, the tail will render 'behind'
the mob and for all other directions, it will render above.

ID cards will now no longer make you look extra pale/bright. The issue
was that the ID card was rendering the body with the parts that've
already got their skin colour/tone, and then applying the skin
colour/tone a second time afterward. No longer an issue.

TO-DO: 'finish' the south dir of certain tail markings (all for vulptail
6, all for default tail, wingler tail)